# improves sort performance