This is just the sort improvements (minus dropping the String() method)

Here is the before and after bench:

```
BenchmarkMatch-8         1000000              1036 ns/op
BenchmarkSort-8          5000000               387 ns/op
PASS
ok      istio.io/istio/pilot/pkg/model  4.930s
```

```
BenchmarkMatch-8        10000000               195 ns/op
BenchmarkSort-8          5000000               318 ns/op
PASS
ok      istio.io/istio/pilot/pkg/model  5.646s
```

The bench is also part of the `service_test` now. If you want to run it on your machine just cd down into the service directory and run `go test bench=.`