@istio-testing 2018-07-25
[APPROVALNOTIFIER] This PR is **NOT APPROVED**

This pull-request has been approved by: *<a href="https://github.com/istio/istio/pull/7399#" title="Author self-approved">zachgersh</a>*
To fully approve this pull request, please assign additional approvers.
We suggest the following additional approver: **kyessenov**

If they are not already assigned, you can assign the PR to them by writing `/assign @kyessenov` in a comment when ready.

The full list of commands accepted by this bot can be found [here](https://go.k8s.io/bot-commands).

The pull request process is described [here](https://git.k8s.io/community/contributors/guide/owners.md#the-code-review-process)

<details open>
Needs approval from an approver in each of these files:

- **[pilot/OWNERS](https://github.com/istio/istio/blob/release-1.0/pilot/OWNERS)**

Approvers can indicate their approval by writing `/approve` in a comment
Approvers can cancel approval by writing `/approve cancel` in a comment
</details>
<!-- META={"approvers":["kyessenov"]} -->